+++ opennms-dao/src/main/java/org/opennms/netmgt/dao/stats/AbstractBaseStatisticsService.java
@@ -45,7 +45,7 @@ public abstract class AbstractBaseStatisticsService<T> implements StatisticsServ
 
     @Transactional
     @Override
-	public int getTotalCount(final Criteria criteria) {
+	public long getTotalCount(final Criteria criteria) {
         return getDao().countMatching(criteria);
 	}
